+++ lib/glimmer/swt/command_handlers/custom_widget_command_handler.rb
@@ -1,5 +1,6 @@
 require File.dirname(__FILE__) + "/../../command_handler"
 require File.dirname(__FILE__) + "/../custom_widget"
+require File.dirname(__FILE__) + "/../custom_shell"
 require File.dirname(__FILE__) + "/../g_widget"
 
 module Glimmer
@@ -9,8 +10,9 @@ module Glimmer
         include CommandHandler
 
         def can_handle?(parent, command_symbol, *args, &block)
-          (parent.is_a?(GWidget) || parent.is_a?(CustomWidget)) and
-          CustomWidget.for(command_symbol)
+          custom_widget_class = CustomWidget.for(command_symbol)
+          custom_widget_class and
+            (parent.is_a?(GWidget) || parent.is_a?(CustomWidget) || custom_widget_class.ancestors.include?(CustomShell))
         end
 
         def do_handle(parent, command_symbol, *args, &block)